+++ doc/guide/build.hpp
@@ -23,14 +23,14 @@ href="https://keon.io/mlpack/mlpack-on-windows/">Keon's excellent tutorial</a>.
 
 @section Download latest mlpack build
 Download latest mlpack build from here:
-<a href="http://www.mlpack.org/files/mlpack-2.1.1.tar.gz">mlpack-2.1.1</a>
+<a href="http://www.mlpack.org/files/mlpack-2.2.0.tar.gz">mlpack-2.2.0</a>
 
 @section builddir Creating Build Directory
 
 Once the mlpack source is unpacked, you should create a build directory.
 
 @code
-$ cd mlpack-2.1.1
+$ cd mlpack-2.2.0
 $ mkdir build
 @endcode
 
@@ -43,8 +43,8 @@ mlpack depends on the following libraries, which need to be installed on the
 system and have headers present:
 
  - Armadillo >= 4.200.0 (with LAPACK support)
- - Boost (math_c99, program_options, serialization, unit_test_framework, heap)
-      >= 1.49
+ - Boost (math_c99, program_options, serialization, unit_test_framework, heap,
+          spirit) >= 1.49
 
 In Ubuntu and Debian, you can get all of these dependencies through apt: