+++ src/mlpack/tests/lars_test.cpp
@@ -279,4 +279,53 @@ BOOST_AUTO_TEST_CASE(RetrainCholeskyTest)
   LARSVerifyCorrectness(betaOpt, errCorr, 0.1);
 }
 
+/**
+ * Make sure that we learn the same when running training separately and through
+ * constructor. Test it with default parameters.
+ */
+BOOST_AUTO_TEST_CASE(TrainingConstructorWithDefaultsTest)
+{
+  arma::mat X;
+  arma::vec y;
+
+  GenerateProblem(X, y, 1000, 100);
+
+  LARS lars1;
+  arma::vec beta;
+  lars1.Train(X, y, beta);
+
+  LARS lars2(X, y);
+
+  BOOST_REQUIRE_EQUAL(beta.n_elem, lars2.BetaPath().back().n_elem);
+  for (size_t i = 0; i < beta.n_elem; ++i)
+    BOOST_REQUIRE_CLOSE(beta[i], lars2.BetaPath().back()[i], 1e-5);
+}
+
+/**
+ * Make sure that we learn the same when running training separately and through
+ * constructor. Test it with non default parameters.
+ */
+BOOST_AUTO_TEST_CASE(TrainingConstructorWithNonDefaultsTest)
+{
+  arma::mat X;
+  arma::vec y;
+
+  GenerateProblem(X, y, 1000, 100);
+
+  bool transposeData = true;
+  bool useCholesky = true;
+  double lambda1 = 0.2;
+  double lambda2 = 0.4;
+
+  LARS lars1(useCholesky, lambda1, lambda2);
+  arma::vec beta;
+  lars1.Train(X, y, beta);
+
+  LARS lars2(X, y, transposeData, useCholesky, lambda1, lambda2);
+
+  BOOST_REQUIRE_EQUAL(beta.n_elem, lars2.BetaPath().back().n_elem);
+  for (size_t i = 0; i < beta.n_elem; ++i)
+    BOOST_REQUIRE_CLOSE(beta[i], lars2.BetaPath().back()[i], 1e-5);
+}
+
 BOOST_AUTO_TEST_SUITE_END();