+++ HISTORY.md
@@ -1,5 +1,8 @@
 ### mlpack ?.?.?
 ###### ????-??-??
+
+### mlpack 2.2.0
+###### 2017-03-21
   * Bugfix for mlpack_knn program (#816).
 
   * Add decision tree implementation in methods/decision_tree/.  This is very
@@ -8,6 +11,12 @@
   * Add DBSCAN implementation in methods/dbscan/.
 
   * Add support for multidimensional discrete distributions (#810, #830).
+
+  * Better output for Log::Debug/Log::Info/Log::Warn/Log::Fatal for Armadillo
+    objects (#895, #928).
+
+  * Refactor categorical CSV loading with boost::spirit for faster loading
+    (#681).
 
 ### mlpack 2.1.1
 ###### 2016-12-22